feat: Add public Empresa Profile
