--- frontend/app/_components/PerfilEmpresa.tsx
@@ -1,142 +1,0 @@
-import { View, Text, Image, TouchableOpacity, StyleSheet } from "react-native";
-import { useAuth } from "../../contexts/AuthContext";
-import colors from "../../assets/styles/colors";
-import { useRouter } from "expo-router";
-import { FontAwesome5, MaterialIcons, Feather } from "@expo/vector-icons";
-
-const PerfilEmpresa = () => {
-    const { user } = useAuth();
-    const router = useRouter();
-
-    return (
-        <View style={styles.container}>
-            <View style={styles.card}>
-                <View style={styles.rowContainer}>
-                    {/* Logo de empresa */}
-                    <View style={styles.profileContainer}>
-                        <Image source={{ uri: user.logo || "https://via.placeholder.com/150" }} style={styles.profileImage} />
-                        {/* Botón de edición */}
-                        <TouchableOpacity style={styles.editIcon} onPress={() => router.push("/miperfil/editar")}>
-                            <Feather name="edit-3" size={22} color={colors.white} />
-                        </TouchableOpacity>
-                    </View>
-                    {/* Información de la empresa */}
-                    <View style={styles.infoContainer}>
-                        <Text style={styles.name}>{user.nombre}</Text>
-                        <Text style={styles.username}>@{user.username}</Text>
-                        <Text style={styles.info}><MaterialIcons name="email" size={18} color={colors.primary} /> {user.email}</Text>
-                        <Text style={styles.info}><MaterialIcons name="phone" size={18} color={colors.primary} /> {user.telefono}</Text>
-                        <Text style={styles.info}><MaterialIcons name="location-pin" size={18} color={colors.primary} /> {user.localizacion}</Text>
-                        <Text style={styles.description}>{user.descripcion}</Text>
-                    </View>
-                </View>
-                {/* Separador */}
-                <View style={styles.separator} />
-
-                <View style={styles.downContainer}>
-                    {/* Información empresarial */}
-                    <Text style={styles.sectionTitle}>Información Empresarial</Text>
-                    <Text style={styles.info}><FontAwesome5 name="id-card" size={18} color={colors.primary} /> NIF: {user.nif}</Text>
-                    <Text style={styles.info}><FontAwesome5 name="globe" size={18} color={colors.primary} /> Web: {user.web}</Text>
-                </View>
-            </View>
-        </View>
-    );
-};
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        alignItems: "center",
-        justifyContent: "center",
-        paddingVertical: 20,
-        backgroundColor: colors.white,
-        marginTop: 20,
-    },
-    card: {
-        backgroundColor: colors.white,
-        padding: 30,
-        borderRadius: 15,
-        elevation: 6,
-        width: "70%",
-        maxWidth: 600,
-        alignSelf: "center",
-        borderWidth: 1,
-        borderColor: colors.lightGray,
-    },
-    rowContainer: {
-        flexDirection: "row",
-        alignItems: "center",
-        justifyContent: "center",
-    },
-    profileContainer: {
-        position: "relative",
-        marginRight: 40,
-    },
-    profileImage: {
-        width: 150,
-        height: 150,
-        borderRadius: 70,
-        borderWidth: 3,
-        borderColor: colors.primary,
-        marginLeft: 30,
-    },
-    editIcon: {
-        position: "absolute",
-        bottom: 10,
-        right: 10,
-        backgroundColor: colors.primary,
-        padding: 8,
-        borderRadius: 15,
-        shadowColor: "#000",
-        shadowOffset: { width: 0, height: 2 },
-        shadowOpacity: 0.2,
-        shadowRadius: 3,
-        elevation: 3,
-    },
-    infoContainer: {
-        flex: 1,
-        justifyContent: "center",
-        marginLeft: 10,
-    },
-    name: {
-        fontSize: 26,
-        fontWeight: "bold",
-        color: colors.secondary,
-    },
-    username: {
-        fontSize: 18,
-        color: colors.darkGray,
-        fontWeight: "bold",
-        marginBottom: 10,
-        marginTop: 6,
-    },
-    info: {
-        fontSize: 16,
-        color: colors.darkGray,
-        marginVertical: 4,
-    },
-    description: {
-        fontSize: 15,
-        color: colors.darkGray,
-        marginTop: 6,
-    },
-    separator: {
-        width: "100%",
-        height: 1,
-        backgroundColor: colors.mediumGray,
-        marginVertical: 20,
-    },
-    sectionTitle: {
-        fontSize: 22,
-        fontWeight: "bold",
-        color: colors.secondary,
-        marginBottom: 15,
-        textAlign: "center",
-    },
-    downContainer: {
-        paddingHorizontal: 30,
-    },
-});
-
-export default PerfilEmpresa;
